Add `setupSchema` method inside `prometheus controller`
